app/backend/requirements.txt
flask
Flask-Cors
mysql-connector-python
python-dotenv
SQLAlchemy
Flask-SQLAlchemy
PyJWT
functools
flask_jwt_extended